# Reduce needless Array generation in some String methods